solved mouse count plots issue
There was a filter for "" in the bin_id col. But mouse one had all NAs,
so this was a edge case. Added a or check for NA now, works. Also added
a new bedfile with the bin_id. Might revert back to the original mm10
bed later. More details, today's note in obsidian.