Major fix for postprocessing filter
The segdups file did not have an entry for every chr, so it was not
finding it in the dt and failed. Added a if check and wrapped it around
the for loop which was giving trouble. Tested, works perfect. Also
passed config as an argument in get_bed function and it works!