make_output_vcf rule works with changes
checked for filtering empty df in add_info to tab file. They were
filtering and adding a value to a col (GT). But the filtering was
resulting in an empty df, which led to the error "replacement has 1 row,
data has 0". To circumvent, I added a check for nrow(subset) > 0. Now
works like a charm. The logic does not change.